remove unwraps from pagerank implementation
